Fix bugs and improve code

+++ web/src/main/java/com/security/web/repositories/ApplicationRepository.java
@@ -4,8 +4,5 @@ import com.security.web.domain.Application;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-//@Transactional(readOnly=true)
 @Repository
-public interface ApplicationRepository extends JpaRepository<Application, Integer> {
-
-}
+public interface ApplicationRepository extends JpaRepository<Application, Long> {}
